# MDP-1754 Improve the memory and cpu performance of reading a subset of columns